+++ CHANGELOG.md
@@ -4,9 +4,10 @@
 
 ### Added
 
-- Added `Get-JiraAgileBoardIssue` for `GET /rest/agile/1.0/board/{boardId}/issue`
-- Added `Get-JiraAgileBacklogIssue` for `GET /rest/agile/1.0/board/{boardId}/backlog`
-- Added `Get-JiraAgileSprintIssue` for `GET /rest/agile/1.0/board/{boardId}/sprint/{sprintId}/issue`
+- Added `Get-JiraAgileIssue` with board, backlog, and sprint parameter sets for:
+  - `GET /rest/agile/1.0/board/{boardId}/issue`
+  - `GET /rest/agile/1.0/board/{boardId}/backlog`
+  - `GET /rest/agile/1.0/board/{boardId}/sprint/{sprintId}/issue`
 - Added `Get-JiraAgileBoardConfiguration` for `GET /rest/agile/1.0/board/{boardId}/configuration`
 - Added `Get-JiraAgileBoardEpic` for `GET /rest/agile/1.0/board/{boardId}/epic`
 - Added `Get-JiraAgileEpic` for `GET /rest/agile/1.0/epic/{epicId}`